Switch to CCL from CXLib
