--- src/main/java/cubex2/mods/morefurnaces/inventory/SlotOutput.java
@@ -1,120 +1,0 @@
-package cubex2.mods.morefurnaces.inventory;
-
-import cubex2.cxlibrary.inventory.ISlotCX;
-import net.minecraft.entity.item.EntityXPOrb;
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.FurnaceRecipes;
-import net.minecraft.util.math.MathHelper;
-import net.minecraftforge.items.IItemHandlerModifiable;
-import net.minecraftforge.items.SlotItemHandler;
-
-import javax.annotation.Nonnull;
-
-public class SlotOutput extends SlotItemHandler implements ISlotCX
-{
-    private final String name;
-    private final EntityPlayer player;
-    private int removeCount;
-
-    public SlotOutput(String name, EntityPlayer player, IItemHandlerModifiable inventoryIn, int slotIndex)
-    {
-        super(inventoryIn, slotIndex, -2000, -2000);
-        this.player = player;
-        this.name = name;
-    }
-
-    @Override
-    public String getName()
-    {
-        return name;
-    }
-
-    /**
-     * Check if the stack is allowed to be placed in this slot, used for armor slots as well as furnace fuel.
-     */
-    public boolean isItemValid(ItemStack stack)
-    {
-        return false;
-    }
-
-    @Override
-    public void onSlotChange(@Nonnull ItemStack p_75220_1_, @Nonnull ItemStack p_75220_2_)
-    {
-        int d = p_75220_2_.getCount() - p_75220_1_.getCount();
-        if (d > 0)
-        {
-            onCrafting(p_75220_2_, d);
-        }
-    }
-
-    /**
-     * Decrease the size of the stack in slot (first int arg) by the amount of the second int arg. Returns the new
-     * stack.
-     */
-    public ItemStack decrStackSize(int amount)
-    {
-        if (this.getHasStack())
-        {
-            this.removeCount += Math.min(amount, this.getStack().getCount());
-        }
-
-        return super.decrStackSize(amount);
-    }
-
-    public ItemStack onTake(EntityPlayer thePlayer, ItemStack stack)
-    {
-        this.onCrafting(stack);
-        super.onTake(thePlayer, stack);
-        return stack;
-    }
-
-    /**
-     * the itemStack passed in is the output - ie, iron ingots, and pickaxes, not ore and wood. Typically increases an
-     * internal count then calls onCrafting(item).
-     */
-    protected void onCrafting(ItemStack stack, int amount)
-    {
-        this.removeCount += amount;
-        this.onCrafting(stack);
-    }
-
-    /**
-     * the itemStack passed in is the output - ie, iron ingots, and pickaxes, not ore and wood.
-     */
-    protected void onCrafting(ItemStack stack)
-    {
-        stack.onCrafting(this.player.world, this.player, this.removeCount);
-
-        if (!this.player.world.isRemote)
-        {
-            int i = this.removeCount;
-            float f = FurnaceRecipes.instance().getSmeltingExperience(stack);
-
-            if (f == 0.0F)
-            {
-                i = 0;
-            } else if (f < 1.0F)
-            {
-                int j = MathHelper.floor((float) i * f);
-
-                if (j < MathHelper.ceil((float) i * f) && Math.random() < (double) ((float) i * f - (float) j))
-                {
-                    ++j;
-                }
-
-                i = j;
-            }
-
-            while (i > 0)
-            {
-                int k = EntityXPOrb.getXPSplit(i);
-                i -= k;
-                this.player.world.spawnEntity(new EntityXPOrb(this.player.world, this.player.posX, this.player.posY + 0.5D, this.player.posZ + 0.5D, k));
-            }
-        }
-
-        this.removeCount = 0;
-        net.minecraftforge.fml.common.FMLCommonHandler.instance().firePlayerSmeltedEvent(player, stack);
-    }
-}
