--- src/main/java/cubex2/mods/morefurnaces/items/ItemMoreFurnaces.java
@@ -1,33 +1,0 @@
-package cubex2.mods.morefurnaces.items;
-
-import cubex2.mods.morefurnaces.FurnaceType;
-import javax.annotation.ParametersAreNonnullByDefault;
-import mcp.MethodsReturnNonnullByDefault;
-import net.minecraft.block.Block;
-import net.minecraft.item.ItemBlock;
-import net.minecraft.item.ItemStack;
-
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-public class ItemMoreFurnaces extends ItemBlock
-{
-
-    public ItemMoreFurnaces(Block block)
-    {
-        super(block);
-        setMaxDamage(0);
-        setHasSubtypes(true);
-    }
-
-    @Override
-    public int getMetadata(int i)
-    {
-        return i;
-    }
-
-    @Override
-    public String getUnlocalizedName(ItemStack stack)
-    {
-        return "tile." + FurnaceType.values()[stack.getItemDamage()].toString().toLowerCase() + "_furnace";
-    }
-}
