--- src/main/java/cubex2/mods/morefurnaces/lib/Textures.java
@@ -1,14 +1,0 @@
-package cubex2.mods.morefurnaces.lib;
-
-import cubex2.cxlibrary.gui.GuiTexture;
-import net.minecraft.util.ResourceLocation;
-
-public class Textures {
-    public static final GuiTexture IRON = new GuiTexture(new ResourceLocation("morefurnaces", "textures/gui/iron.png"), 256, 256, true);
-    public static final GuiTexture GOLD = new GuiTexture(new ResourceLocation("morefurnaces", "textures/gui/gold.png"), 256, 256, true);
-    public static final GuiTexture DIAMOND = new GuiTexture(new ResourceLocation("morefurnaces", "textures/gui/diamond.png"), 256, 256, true);
-    public static final GuiTexture NETHERRACK = new GuiTexture(new ResourceLocation("morefurnaces", "textures/gui/netherrack.png"), 256, 256, true);
-    public static final GuiTexture OBSIDIAN = new GuiTexture(new ResourceLocation("morefurnaces", "textures/gui/obsidian.png"), 256, 256, true);
-    public static final GuiTexture COPPER = new GuiTexture(new ResourceLocation("morefurnaces", "textures/gui/copper.png"), 256, 256, true);
-    public static final GuiTexture SILVER = new GuiTexture(new ResourceLocation("morefurnaces", "textures/gui/silver.png"), 256, 256, true);
-}
