--- src/main/java/cubex2/mods/morefurnaces/tileentity/TileEntityCopperFurnace.java
@@ -1,11 +1,0 @@
-package cubex2.mods.morefurnaces.tileentity;
-
-import cubex2.mods.morefurnaces.FurnaceType;
-
-public class TileEntityCopperFurnace extends TileEntityIronFurnace
-{
-    public TileEntityCopperFurnace()
-    {
-        super(FurnaceType.COPPER);
-    }
-}
