--- src/main/java/cubex2/mods/morefurnaces/tileentity/TileEntityDiamondFurnace.java
@@ -1,11 +1,0 @@
-package cubex2.mods.morefurnaces.tileentity;
-
-import cubex2.mods.morefurnaces.FurnaceType;
-
-public class TileEntityDiamondFurnace extends TileEntityIronFurnace
-{
-    public TileEntityDiamondFurnace()
-    {
-        super(FurnaceType.DIAMOND);
-    }
-}
