--- src/main/java/cubex2/mods/morefurnaces/tileentity/TileEntityGoldFurnace.java
@@ -1,9 +1,0 @@
-package cubex2.mods.morefurnaces.tileentity;
-
-import cubex2.mods.morefurnaces.FurnaceType;
-
-public class TileEntityGoldFurnace extends TileEntityIronFurnace {
-    public TileEntityGoldFurnace() {
-        super(FurnaceType.GOLD);
-    }
-}
