--- src/main/java/cubex2/mods/morefurnaces/tileentity/TileEntityObsidianFurnace.java
@@ -1,11 +1,0 @@
-package cubex2.mods.morefurnaces.tileentity;
-
-import cubex2.mods.morefurnaces.FurnaceType;
-
-public class TileEntityObsidianFurnace extends TileEntityIronFurnace
-{
-    public TileEntityObsidianFurnace()
-    {
-        super(FurnaceType.OBSIDIAN);
-    }
-}
