--- src/main/java/cubex2/mods/morefurnaces/tileentity/TileEntitySilverFurnace.java
@@ -1,11 +1,0 @@
-package cubex2.mods.morefurnaces.tileentity;
-
-import cubex2.mods.morefurnaces.FurnaceType;
-
-public class TileEntitySilverFurnace extends TileEntityIronFurnace
-{
-    public TileEntitySilverFurnace()
-    {
-        super(FurnaceType.SILVER);
-    }
-}
