+++ src/main/java/dan/morefurnaces/tileentity/TileEntityCopperFurnace.java
@@ -1,0 +1,9 @@
+package dan.morefurnaces.tileentity;
+
+import dan.morefurnaces.FurnaceType;
+
+public class TileEntityCopperFurnace extends TileEntityIronFurnace {
+    public TileEntityCopperFurnace() {
+        super(FurnaceType.COPPER);
+    }
+}
