+++ src/main/java/dan/morefurnaces/tileentity/TileEntityDiamondFurnace.java
@@ -1,0 +1,9 @@
+package dan.morefurnaces.tileentity;
+
+import dan.morefurnaces.FurnaceType;
+
+public class TileEntityDiamondFurnace extends TileEntityIronFurnace {
+    public TileEntityDiamondFurnace() {
+        super(FurnaceType.DIAMOND);
+    }
+}
