+++ src/main/java/dan/morefurnaces/tileentity/TileEntityGoldFurnace.java
@@ -1,0 +1,9 @@
+package dan.morefurnaces.tileentity;
+
+import dan.morefurnaces.FurnaceType;
+
+public class TileEntityGoldFurnace extends TileEntityIronFurnace {
+    public TileEntityGoldFurnace() {
+        super(FurnaceType.GOLD);
+    }
+}
