+++ src/main/java/dan/morefurnaces/tileentity/TileEntityObsidianFurnace.java
@@ -1,0 +1,9 @@
+package dan.morefurnaces.tileentity;
+
+import dan.morefurnaces.FurnaceType;
+
+public class TileEntityObsidianFurnace extends TileEntityIronFurnace {
+    public TileEntityObsidianFurnace() {
+        super(FurnaceType.OBSIDIAN);
+    }
+}
