+++ src/main/java/dan/morefurnaces/tileentity/TileEntitySilverFurnace.java
@@ -1,0 +1,9 @@
+package dan.morefurnaces.tileentity;
+
+import dan.morefurnaces.FurnaceType;
+
+public class TileEntitySilverFurnace extends TileEntityIronFurnace {
+    public TileEntitySilverFurnace() {
+        super(FurnaceType.SILVER);
+    }
+}
