+++ src/main/resources/assets/morefurnaces/gui/gold.json
@@ -2,30 +2,44 @@
   "slots": {
     "player": [
       {
-        "from": 0, "to": 8,
-        "x": 8, "y": 142
+        "from": 0,
+        "to": 8,
+        "x": 8,
+        "y": 142
       },
       {
-        "from": 9, "to": 35, "rows": 3,
-        "x": 8, "y": 84
+        "from": 9,
+        "to": 35,
+        "rows": 3,
+        "x": 8,
+        "y": 84
       }
     ],
     "furnace": [
       {
-        "from": 3, "to": 0,
-        "x": 8, "y": 17
+        "from": 3,
+        "to": 0,
+        "x": 8,
+        "y": 17
       },
       {
-        "from": 7, "to": 4,
-        "x": 8, "y": 53
+        "from": 7,
+        "to": 4,
+        "x": 8,
+        "y": 53
       },
       {
-        "from": 8, "to": 8,
-        "x": 112, "y": 35
+        "from": 8,
+        "to": 8,
+        "x": 112,
+        "y": 35
       },
       {
-        "from": 9, "to": 12, "rows": 2,
-        "x": 134, "y": 26
+        "from": 9,
+        "to": 12,
+        "rows": 2,
+        "x": 134,
+        "y": 26
       }
     ]
   },
