+++ src/main/resources/assets/morefurnaces/gui/netherrack.json
@@ -2,22 +2,31 @@
   "slots": {
     "player": [
       {
-        "from": 0, "to": 8,
-        "x": 8, "y": 142
+        "from": 0,
+        "to": 8,
+        "x": 8,
+        "y": 142
       },
       {
-        "from": 9, "to": 35, "rows": 3,
-        "x": 8, "y": 84
+        "from": 9,
+        "to": 35,
+        "rows": 3,
+        "x": 8,
+        "y": 84
       }
     ],
     "furnace": [
       {
-        "from": 0, "to": 0,
-        "x": 56, "y": 17
+        "from": 0,
+        "to": 0,
+        "x": 56,
+        "y": 17
       },
       {
-        "from": 1, "to": 1,
-        "x": 116, "y": 35
+        "from": 1,
+        "to": 1,
+        "x": 116,
+        "y": 35
       }
     ]
   },
