+++ src/main/resources/assets/morefurnaces/gui/obsidian.json
@@ -2,61 +2,87 @@
   "slots": {
     "player": [
       {
-        "from": 0, "to": 8,
-        "x": 8, "y": 172
+        "from": 0,
+        "to": 8,
+        "x": 8,
+        "y": 172
       },
       {
-        "from": 9, "to": 35, "rows": 3,
-        "x": 8, "y": 114
+        "from": 9,
+        "to": 35,
+        "rows": 3,
+        "x": 8,
+        "y": 114
       }
     ],
     "furnace": [
       {
-        "from": 1, "to": 0,
-        "x": 38, "y": 17
+        "from": 1,
+        "to": 0,
+        "x": 38,
+        "y": 17
       },
       {
-        "from": 3, "to": 2,
-        "x": 38, "y": 43
+        "from": 3,
+        "to": 2,
+        "x": 38,
+        "y": 43
       },
       {
-        "from": 5, "to": 4,
-        "x": 38, "y": 83
+        "from": 5,
+        "to": 4,
+        "x": 38,
+        "y": 83
       },
       {
-        "from": 6, "to": 6,
-        "x": 116, "y": 18
+        "from": 6,
+        "to": 6,
+        "x": 116,
+        "y": 18
       },
       {
-        "from": 7, "to": 7,
-        "x": 138, "y": 22
+        "from": 7,
+        "to": 7,
+        "x": 138,
+        "y": 22
       },
       {
-        "from": 8, "to": 8,
-        "x": 116, "y": 44
+        "from": 8,
+        "to": 8,
+        "x": 116,
+        "y": 44
       },
       {
-        "from": 9, "to": 9,
-        "x": 138, "y": 48
+        "from": 9,
+        "to": 9,
+        "x": 138,
+        "y": 48
       }
     ]
   },
   "controls": {
     "bg": {
-      "width": 176, "height": 196,
+      "width": 176,
+      "height": 196,
       "center": true
     },
     "cook0": {
-      "width": 24, "height": 16,
-      "left": 79, "top": 17
+      "width": 24,
+      "height": 16,
+      "left": 79,
+      "top": 17
     },
     "cook1": {
-      "width": 24, "height": 16,
-      "left": 79, "top": 43
+      "width": 24,
+      "height": 16,
+      "left": 79,
+      "top": 43
     },
     "fuel": {
-      "width": 14, "height": 13,
-      "left": 56, "top": 66
+      "width": 14,
+      "height": 13,
+      "left": 56,
+      "top": 66
     }
   }
 }
