+++ src/main/resources/assets/morefurnaces/models/item/furnace_obsidian.json
@@ -2,9 +2,21 @@
   "parent": "morefurnaces:block/furnace_obsidian",
   "display": {
     "thirdperson": {
-      "rotation": [ 10, -45, 170 ],
-      "translation": [ 0, 1.5, -2.75 ],
-      "scale": [ 0.375, 0.375, 0.375 ]
+      "rotation": [
+        10,
+        -45,
+        170
+      ],
+      "translation": [
+        0,
+        1.5,
+        -2.75
+      ],
+      "scale": [
+        0.375,
+        0.375,
+        0.375
+      ]
     }
   }
 }
